requirements.txt
openai
chromadb
pypdf
langchain
wikipedia
tavily-python